Renamed Measurement* to MeasurementRecording* to avoid confusion
