+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/ScreensModule.kt
@@ -5,8 +5,8 @@ import org.koin.dsl.module
 import org.noiseplanet.noisecapture.ui.features.details.MeasurementDetailsScreenViewModel
 import org.noiseplanet.noisecapture.ui.features.history.HistoryScreenViewModel
 import org.noiseplanet.noisecapture.ui.features.home.HomeScreenViewModel
-import org.noiseplanet.noisecapture.ui.features.measurement.MeasurementScreenViewModel
 import org.noiseplanet.noisecapture.ui.features.permission.RequestPermissionScreenViewModel
+import org.noiseplanet.noisecapture.ui.features.recording.MeasurementRecordingScreenViewModel
 import org.noiseplanet.noisecapture.ui.features.settings.SettingsScreenViewModel
 
 val screensModule = module {
@@ -18,7 +18,7 @@ val screensModule = module {
     }
 
     viewModel {
-        MeasurementScreenViewModel()
+        MeasurementRecordingScreenViewModel()
     }
 
     viewModel { (measurementId: String) ->
