+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/recording/MeasurementRecordingModule.kt
@@ -1,14 +1,14 @@
-package org.noiseplanet.noisecapture.ui.features.measurement
+package org.noiseplanet.noisecapture.ui.features.recording
 
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 import org.noiseplanet.noisecapture.ui.components.spl.SoundLevelMeterViewModel
-import org.noiseplanet.noisecapture.ui.features.measurement.controls.RecordingControlsViewModel
-import org.noiseplanet.noisecapture.ui.features.measurement.plot.spectrogram.SpectrogramPlotViewModel
-import org.noiseplanet.noisecapture.ui.features.measurement.plot.spectrum.SpectrumPlotViewModel
+import org.noiseplanet.noisecapture.ui.features.recording.controls.RecordingControlsViewModel
+import org.noiseplanet.noisecapture.ui.features.recording.plot.spectrogram.SpectrogramPlotViewModel
+import org.noiseplanet.noisecapture.ui.features.recording.plot.spectrum.SpectrumPlotViewModel
 
 
-val measurementModule = module {
+val measurementRecordingModule = module {
 
     viewModel {
         SoundLevelMeterViewModel(
