+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/recording/MeasurementRecordingPager.kt
@@ -1,4 +1,4 @@
-package org.noiseplanet.noisecapture.ui.features.measurement
+package org.noiseplanet.noisecapture.ui.features.recording
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -15,15 +15,15 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import kotlinx.coroutines.launch
 import org.koin.compose.viewmodel.koinViewModel
-import org.noiseplanet.noisecapture.ui.features.measurement.plot.spectrogram.SpectrogramPlotView
-import org.noiseplanet.noisecapture.ui.features.measurement.plot.spectrum.SpectrumPlotView
+import org.noiseplanet.noisecapture.ui.features.recording.plot.spectrogram.SpectrogramPlotView
+import org.noiseplanet.noisecapture.ui.features.recording.plot.spectrum.SpectrumPlotView
 
 /**
  * A horizontal pager on the measurement recording screens that allows user to switch between
  * spectrum, spectrogram and map views.
  */
 @Composable
-fun MeasurementPager(
+fun MeasurementRecordingPager(
     modifier: Modifier = Modifier,
 ) {
