+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/recording/controls/RecordingControls.kt
@@ -1,4 +1,4 @@
-package org.noiseplanet.noisecapture.ui.features.measurement.controls
+package org.noiseplanet.noisecapture.ui.features.recording.controls
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandHorizontally
