+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/recording/plot/LegendElement.kt
@@ -1,4 +1,4 @@
-package org.noiseplanet.noisecapture.ui.features.measurement.plot
+package org.noiseplanet.noisecapture.ui.features.recording.plot
 
 import androidx.compose.ui.text.TextLayoutResult
