+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/recording/plot/spectrogram/SpectrogramPlotViewModel.kt
@@ -1,4 +1,4 @@
-package org.noiseplanet.noisecapture.ui.features.measurement.plot.spectrogram
+package org.noiseplanet.noisecapture.ui.features.recording.plot.spectrogram
 
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.ui.unit.IntSize
